+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.0.1 under development
 
-- no changes in this release.
+- Bug #1127: Fix `AbstractSchema::hasTable()` and `AbstractSchema::hasView()` methods to support names quoted with curly
+  brackets `{{%table}}` (@batyrmastyr)
 
 ## 2.0.0 December 05, 2025